+++ CHANGELOG.md
@@ -5,6 +5,29 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [0.6.0](https://github.com/crowdsecurity/php-remediation-engine/releases/tag/v0.6.0) - 2023-01-12
+[_Compare with previous release_](https://github.com/crowdsecurity/php-remediation-engine/compare/v0.5.0...v0.6.0)
+
+### Changed
+
+- For LAPI in live mode, we use and save retrieved IPv6 range scoped decisions as IP scoped decisions 
+- Unexpected configuration keys are automatically removed by a new `cleanConfigs` method
+- Do not try to retrieve range scoped decisions in cache for IPv6 IP as it is not yet implemented
+- Update some logs
+- Update `crowdsec/capi-client` dependency to `v0.10.0`
+- Update `crowdsec/lapi-client` dependency to `v0.4.0`
+- Do not throw exception for unknown prefix in `getCacheKey` method
+
+### Added
+
+- Add public method `getCacheStorage` for remediations
+- Add public method `unsetIpVariables` for cache
+- Add public method `clearGeolocationCache` for geolocation
+
+--- 
+
+
+
 ## [0.5.0](https://github.com/crowdsecurity/php-remediation-engine/releases/tag/v0.5.0) - 2023-01-05
 [_Compare with previous release_](https://github.com/crowdsecurity/php-remediation-engine/compare/v0.4.0...v0.5.0)